feat: update schedule types
